requirements.txt
torch==2.4.0
transformers==4.38.0
sentence-transformers==2.5.0
datasets==2.18.0
pyyaml==6.0.1
numpy==1.26.4
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2
wandb==0.16.4
einops==0.7.0
evaluate==0.4.1
bert-score==0.3.13
bitsandbytes==0.42.0
accelerate==0.27.0
rouge_score
nltk
absl-py
torch_xla
